server/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
psycopg2-binary==2.8.6
python-dotenv==0.15.0
